+++ host/class/uac/usb_host_uac/include/usb/uac_host.h
@@ -381,7 +381,7 @@ esp_err_t uac_host_device_write(uac_host_device_handle_t uac_dev_handle, uint8_t
 
 /**
  * @brief Mute or un-mute the UAC device
- * @param[in] iface       Pointer to UAC interface structure
+ * @param[in] uac_dev_handle  UAC device handle
  * @param[in] mute        True to mute, false to unmute
  * @return esp_err_t
  * - ESP_OK on success
@@ -393,8 +393,21 @@ esp_err_t uac_host_device_write(uac_host_device_handle_t uac_dev_handle, uint8_t
 esp_err_t uac_host_device_set_mute(uac_host_device_handle_t uac_dev_handle, bool mute);
 
 /**
+ * @brief Get the mute status of the UAC device
+ * @param[in] uac_dev_handle  UAC device handle
+ * @param[out] mute       Pointer to store the mute status
+ * @return esp_err_t
+ * - ESP_OK on success
+ * - ESP_ERR_INVALID_STATE if the device is not ready or active
+ * - ESP_ERR_INVALID_ARG if the device handle is invalid
+ * - ESP_ERR_NOT_SUPPORTED if the device does not support mute control
+ * - ESP_ERR_TIMEOUT if the control timed out
+ */
+esp_err_t uac_host_device_get_mute(uac_host_device_handle_t uac_dev_handle, bool *mute);
+
+/**
  * @brief Set the volume of the UAC device
- * @param[in] iface       Pointer to UAC interface structure
+ * @param[in] uac_dev_handle  UAC device handle
  * @param[in] volume      Volume to set, 0-100
  * @return esp_err_t
  * - ESP_OK on success
@@ -406,9 +419,9 @@ esp_err_t uac_host_device_set_mute(uac_host_device_handle_t uac_dev_handle, bool
 esp_err_t uac_host_device_set_volume(uac_host_device_handle_t uac_dev_handle, uint8_t volume);
 
 /**
- * @brief Set the volume of the UAC device in dB
- * @param[in] iface       Pointer to UAC interface structure
- * @param[in] volume_db   Volume to set, db
+ * @brief Get the volume of the UAC device
+ * @param[in] uac_dev_handle  UAC device handle
+ * @param[out] volume     Pointer to store the volume, 0-100
  * @return esp_err_t
  * - ESP_OK on success
  * - ESP_ERR_INVALID_STATE if the device is not ready or active
@@ -416,7 +429,35 @@ esp_err_t uac_host_device_set_volume(uac_host_device_handle_t uac_dev_handle, ui
  * - ESP_ERR_NOT_SUPPORTED if the device does not support volume control
  * - ESP_ERR_TIMEOUT if the control timed out
  */
-esp_err_t uac_host_device_set_volume_db(uac_host_device_handle_t uac_dev_handle, uint32_t volume_db);
+esp_err_t uac_host_device_get_volume(uac_host_device_handle_t uac_dev_handle, uint8_t *volume);
+
+/**
+ * @brief Set the volume of the UAC device in dB
+ * @param[in] uac_dev_handle  UAC device handle
+ * @param[in] volume_db   Volume to set, with resolution of 1/256 dB,
+ * eg.  256 (0x0100) is 1 dB. 32767 (0x7FFF) is 127.996 dB. -32767 (0x8001) is -127.996 dB.
+ * @return esp_err_t
+ * - ESP_OK on success
+ * - ESP_ERR_INVALID_STATE if the device is not ready or active
+ * - ESP_ERR_INVALID_ARG if the device handle is invalid
+ * - ESP_ERR_NOT_SUPPORTED if the device does not support volume control
+ * - ESP_ERR_TIMEOUT if the control timed out
+ */
+esp_err_t uac_host_device_set_volume_db(uac_host_device_handle_t uac_dev_handle, int16_t volume_db);
+
+/**
+ * @brief Get the volume of the UAC device in dB
+ * @param[in] uac_dev_handle  UAC device handle
+ * @param[out] volume_db  Pointer to store the volume, with resolution of 1/256 dB,
+ * eg.  256 (0x0100) is 1 dB. 32767 (0x7FFF) is 127.996 dB. -32767 (0x8001) is -127.996 dB.
+ * @return esp_err_t
+ * - ESP_OK on success
+ * - ESP_ERR_INVALID_STATE if the device is not ready or active
+ * - ESP_ERR_INVALID_ARG if the device handle is invalid
+ * - ESP_ERR_NOT_SUPPORTED if the device does not support volume control
+ * - ESP_ERR_TIMEOUT if the control timed out
+ */
+esp_err_t uac_host_device_get_volume_db(uac_host_device_handle_t uac_dev_handle, int16_t *volume_db);
 
 #ifdef __cplusplus
 }