+++ host/class/uac/usb_host_uac/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog for USB Host UAC
 
+## 1.2.0 2024-09-27
+
+### Breaking Changes:
+
+1. Changed the parameter type of `uac_host_device_set_volume_db` from uint32_t to int16_t
+
+
+### Improvements:
+
+1. Support get current volume and mute status
+
+### Bugfixes:
+
+1. Fixed incorrect volume conversion. Using actual device volume range.
+2. Fixed concurrency issues when suspend/stop during read/write
+
 ## 1.1.0
 
 ### Improvements